+++ .env
@@ -1,7 +1,2 @@
 ESLINT_NO_DEV_ERRORS=true
 TSC_COMPILE_ON_ERROR=true
-MONITOR_BACKEND_USE_LOCATION=false
-MONITOR_BACKEND_PATH_SOCKETIO=/hkube/monitor-server/socket.io
-MONITOR_BACKEND_HOST=dev1.hkube.io
-MONITOR_BACKEND_PORT=80
-MONITOR_BACKEND_PATH=/hkube/monitor-server
