+++ src/hooks/usePipeline.js
@@ -4,6 +4,7 @@ import { selectors } from 'reducers';
 import { message } from 'antd';
 import client from 'client';
 import successMsg from 'config/schema/success-messages.schema';
+import { useHistory } from 'react-router-dom';
 import useActions from './useActions';
 import { useFilter } from './useSearch';
 
@@ -13,6 +14,8 @@ const usePipeline = () => {
   const dataStats = useSelector(selectors.pipelines.stats.all);
   const filtered = useFilter(collection, ['name']);
   const { updateStored } = useActions();
+  const history = useHistory();
+
   const updateCron = useCallback(
     (pipeline, pattern) => {
       updateStored({
@@ -26,7 +29,7 @@ const usePipeline = () => {
     [updateStored]
   );
 
-  const rerunPipeline = useCallback(async (jobId) => {
+  const rerunPipeline = useCallback(async jobId => {
     try {
       const res = await client.post('/exec/rerun', { jobId });
       message.success(successMsg(res.data).PIPELINE_START);
@@ -35,11 +38,23 @@ const usePipeline = () => {
     }
   }, []);
 
+  const addNewPipeline = useCallback(async (data, LOCAL_STORAGE_KEY) => {
+    try {
+      const res = await client.post('store/pipelines', { ...data });
+      message.success(successMsg(res.data).PIPELINE_ADD);
+      window.localStorage.removeItem(LOCAL_STORAGE_KEY);
+      history.push('/pipelines');
+    } catch (res) {
+      message.error(res.response.data.error.message);
+    }
+  }, []);
+
   return {
     collection: filtered,
     dataStats,
     updateCron,
     rerunPipeline,
+    addNewPipeline,
   };
 };
